Add new components for RealtimeStepSummary: AffectedRides, AlertBasicInfo, and CauseAndEffect. Refactor index.tsx to integrate these components, enhancing the structure and readability of the summary section.
